refactor: provider enum 모듈 이동
