+++ soridam-auth/src/main/java/sorisoop/soridam/auth/oauth/OidcService.java
@@ -15,7 +15,7 @@ import sorisoop.soridam.auth.oauth.exception.OidcInvalidAudienceException;
 import sorisoop.soridam.auth.oauth.exception.OidcInvalidIdTokenException;
 import sorisoop.soridam.auth.oauth.exception.OidcInvalidIssuerException;
 import sorisoop.soridam.auth.oauth.request.OidcLoginRequest;
-import sorisoop.soridam.domain.user.domain.Provider;
+import sorisoop.soridam.common.domain.Provider;
 import sorisoop.soridam.domain.user.domain.User;
 import sorisoop.soridam.domain.user.infrastructure.JpaUserRepository;
