+++ soridam-common/src/main/java/sorisoop/soridam/common/exception/ProviderNotFoundException.java
@@ -1,0 +1,9 @@
+package sorisoop.soridam.common.exception;
+
+import static sorisoop.soridam.common.exception.GlobalExceptionCode.PROVIDER_NOT_FOUND;
+
+public class ProviderNotFoundException extends CustomException {
+	public ProviderNotFoundException() {
+		super(PROVIDER_NOT_FOUND);
+	}
+}
