+++ soridam-domain/src/main/java/sorisoop/soridam/domain/user/infrastructure/JpaUserRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import sorisoop.soridam.domain.user.domain.Provider;
+import sorisoop.soridam.common.domain.Provider;
 import sorisoop.soridam.domain.user.domain.User;
 
 public interface JpaUserRepository extends JpaRepository<User, String> {
